backend/requirements.txt
fastapi==0.115.0
uvicorn==0.35.0
pydantic==2.9.2
python-multipart==0.0.12
pypdf==4.3.1
jinja2==3.1.4
httpx==0.28.1
cohere
numpy
scikit-learn
python-dotenv

tqdm

langchain

python-docx
